Moved scope source code to its separate folder
